+++ Tests/VCoreTests/Tests/Extensions/Foundation/Sequence/SequenceSortByKeyPathTests.swift
@@ -1,5 +1,5 @@
 //
-//  SequenceKeyPathSortTests.swift
+//  SequenceSortByKeyPathTests.swift
 //  VCoreTests
 //
 //  Created by Vakhtang Kontridze on 06.05.22.
@@ -9,7 +9,7 @@ import XCTest
 @testable import VCore
 
 // MARK: - Tests
-final class SequenceKeyPathSortTests: XCTestCase {
+final class SequenceSortByKeyPathTests: XCTestCase {
     // MARK: Test Data
     private struct City: Equatable {
         let name: String